ids from kramerius

+++ kramids.sh
@@ -4,8 +4,20 @@ source="$1"
 i=0
 count=5000
 
+if [ -z $source ]; then
+  printf "No source - mzk | nkp \n"
+  exit 1
+fi
+source="kram-$1"
+
+sourceUrl=$(bash ini.sh getValue $source)
 logPath=$(bash ini.sh getValue defaultLogPath)
-dataPath=$(bash ini.sh getValue defaultDataPath)'kram'
+dataPath=$(bash ini.sh getValue defaultDataPath)$source
+
+if [ -z $sourceUrl ]; then
+  printf "No url for $source \n"
+  exit 1
+fi
 
 if [ -d "$dataPath" ]; then
   rm $dataPath/*
@@ -13,7 +25,7 @@ else
   mkdir $dataPath
 fi
 
-url="https://$source/search/api/v5.0/search?q=*:*&fq=fedora.model:monograph+OR+fedora.model:periodical+OR+fedora.model:map+OR+fedora.model:archive+OR+fedora.model:manuscript+OR+fedora.model:sheetmusic+OR+fedora.model:soundrecording+OR+fedora.model:graphic&fl=PID&rows=$count&start="
+url="$sourceUrl/search/api/v5.0/search?q=*:*&fq=fedora.model:monograph+OR+fedora.model:periodical+OR+fedora.model:map+OR+fedora.model:archive+OR+fedora.model:manuscript+OR+fedora.model:sheetmusic+OR+fedora.model:soundrecording+OR+fedora.model:graphic&fl=PID&rows=$count&start="
 now=$(date)
 error=0
 echo "$now: Downloading URL: $url$i" >$logPath'kram.log'
